Introduction about the License Management System
The license management system plays a crucial role in digitizing traditional license management techniques in a digital way. This system is designed in such a way that both technical and non-technical people can interact with it. The main motive of this system was to make the license registration, application, review, approval, and reporting process simple and faster. This system also helps in license data management for the government to use for research purposes. We can also expand the feature of this system, as we were suggesting to implement monitoring real-time traffic violations, with the help of which the system will automatically debit the traffic fine from the user’s bank account, which will be done only in a month. We also use many verification techniques for security, like criteria for password creation or uploading original documents for verification, etc. In today’s rules of government are changing day by day, so it also focuses on updating user’s when the government updates the rules of traffic, which will alert drivers to traffic violations.